add one log to solve the http request stuck issue
Details: when I run the script to update all the stocks, sometimes, the process will be stuck even almost all the stocks have been processed. Strangely, I did not see any log in the log file. It only sent out one http request, but nothing returned. I want to debug for this, so I added this log. But the problem disappeared itself.
I guess the reason may be: the request has been sent out, but during waiting for the data one issue occurred. And the node will wait for timeout. But the default timeout may be too long. This is not verified. I could verify it when I have time. And for the timeout parameter, I could refer to: http://nodejs.cn/api/http.html#http_http_request_options_callback